refactored code by separating all functionality/utility for looking up abilities into a separate component
